pyproject.toml
[project]
name = "open-log-less"
version = "0.1.0"
description = "A open source log api"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "djlint>=1.36.4",
    "fastapi[standard]==0.99.0",
    "jinja2>=3.1.6",
    "psycopg>=3.3.1",
    "pydantic<2",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]


[project.scripts]
opls="app.db.migrate:create_tables"
